+++ index.html
@@ -27,7 +27,6 @@
         ></a>
       </div>
     </div>
-    <div class="text">lorem</div>
     <!-- <div class="quizContainer">
       <h3>Question1</h3>
       <div class="question m-5">
