+++ src/views/footer.view.js
@@ -1,0 +1,12 @@
+export const createFooter = (currentQuestion, totalQuestions) => {
+  const quizFooter = document.createElement('div');
+  quizFooter.classList.add('quizFooter');
+  quizFooter.innerHTML = `<a id="prevButton" href="#" class="prevBtn"
+  ><i class="fa fa-arrow-left prevArrow"></i> Pervious</a
+>
+<div class="quizFooterText">${currentQuestion}/${totalQuestions}</div>
+<a id="nextButton" href="#" class="nextBtn"
+  >Next <i class="fa fa-arrow-right nextArrow"></i
+></a>`;
+  return quizFooter;
+};
